app/schemas.py
from pydantic import BaseModel


class UserCreate(BaseModel):
    name: str
    email: str
    role: str



class UserLogin(BaseModel):
    email: str
    password: str


class Token(BaseModel):
    access_token: str
    token_type: str = "bearer"


from pydantic import BaseModel, EmailStr

class LoginRequest(BaseModel):
    email: EmailStr
    password: str
